Realize Doran is stupid and we need to seperate out the "tinker with X" objects into different classes.
The plan is that there is a "Judge" class for each type of component to be randomized (Champion, Item, SummonerSpell). It takes a just "bland old blob" from DataDragon on init, and uses an internal state to compose and precompute different datapoints (like groupings for items) based on the config file to allow for the Javascript side to flow smoother.

ItemJudge also seperates out different operations that I was doing in Tristana/Doran to be more sane.